requirements.txt
Cython==0.13
http://www.djangoproject.com/download/1.3-beta-1/tarball/
amqplib==0.6.1
anyjson==0.3
carrot==0.10.7
celery==2.1.4
distribute==0.6.14
django-celery==2.1.4
django-indexer==0.2.1
django-paging==0.2.2
django-picklefield==0.1.9
django-registration==0.7
django-sentry==1.5.3
django-templatetag-sugar==0.1
gunicorn==0.12.0
pyparsing==1.5.5
python-dateutil==1.5
urwid==0.9.9.1
uuid==1.30
virtualenv==1.5.1
wsgiref==0.1.2
